+++ server/common/src/test/kotlin/com/kenlikdev/qmarket/common/security/JwtServiceTest.kt
@@ -6,7 +6,6 @@ import org.junit.jupiter.api.Assertions.assertThrows
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
-import java.util.List
 import java.util.UUID
 
 class JwtServiceTest {